+++ djstripe/models.py
@@ -5,6 +5,7 @@ import datetime
 import decimal
 import json
 import traceback as exception_traceback
+import warnings
 
 from django.conf import settings
 from django.contrib.sites.models import Site
@@ -12,7 +13,7 @@ from django.core.mail import EmailMessage
 from django.db import models
 from django.template.loader import render_to_string
 from django.utils import timezone
-from django.utils.encoding import python_2_unicode_compatible
+from django.utils.encoding import python_2_unicode_compatible, smart_text
 
 from jsonfield.fields import JSONField
 from model_utils.models import TimeStampedModel
@@ -28,10 +29,6 @@ from .signals import webhook_processing_error
 
 stripe.api_key = settings.STRIPE_SECRET_KEY
 stripe.api_version = getattr(settings, "STRIPE_API_VERSION", "2014-01-31")
-
-
-if djstripe_settings.PY3:
-    unicode = str
 
 
 def convert_tstamp(response, field_name=None):
@@ -71,7 +68,7 @@ class EventProcessingException(TimeStampedModel):
         )
 
     def __str__(self):
-        return "<%s, pk=%s, Event=%s>" % (self.message, self.pk, self.event)
+        return "<{message}, pk={pk}, Event={event}>".format(message=self.message, pk=self.pk, event=self.event)
 
 
 @python_2_unicode_compatible
@@ -90,7 +87,7 @@ class Event(StripeObject):
         return self.validated_message
 
     def __str__(self):
-        return "%s - %s" % (self.kind, self.stripe_id)
+        return "<{kind}, stripe_id={stripe_id}>".format(kind=self.kind, stripe_id=self.stripe_id)
 
     def link_customer(self):
         stripe_customer_id = None
@@ -120,11 +117,7 @@ class Event(StripeObject):
                 cls=stripe.StripeObjectEncoder
             )
         )
-        if self.webhook_message["data"] == self.validated_message["data"]:
-            self.valid = True
-        else:
-            # TODO - needs test
-            self.valid = False
+        self.valid = self.webhook_message["data"] == self.validated_message["data"]
         self.save()
 
     def process(self):
@@ -167,58 +160,44 @@ class Event(StripeObject):
             "ping"
         """
         if self.valid and not self.processed:
-            # TODO - needs tests
             try:
-                if not self.kind.startswith("plan.") and \
-                        not self.kind.startswith("transfer."):
+                # Link the customer
+                if not self.kind.startswith("plan.") and not self.kind.startswith("transfer."):
                     self.link_customer()
+
+                # Handle events
                 if self.kind.startswith("invoice."):
                     Invoice.handle_event(self)
                 elif self.kind.startswith("charge."):
-                    if not self.customer:
-                        self.link_customer()
-                    self.customer.record_charge(
-                        self.message["data"]["object"]["id"]
-                    )
+                    self.customer.record_charge(self.message["data"]["object"]["id"])
                 elif self.kind.startswith("transfer."):
-                    Transfer.process_transfer(
-                        self,
-                        self.message["data"]["object"]
-                    )
+                    Transfer.process_transfer(self, self.message["data"]["object"])
                 elif self.kind.startswith("customer.subscription."):
-                    if not self.customer:
-                        self.link_customer()
                     if self.customer:
-                        # TODO see whether sync_subscriptions() also covers the deletion case.
-                        """
-                        Previously, the code tested for "customer.subscription.deleted" and in
-                        this case would cancel the current_subscription, rather than syncing.
-
-                        if self.kind == "customer.subscription.deleted":
-                            self.customer.current_subscription.status = CurrentSubscription.STATUS_CANCELLED
-                            self.customer.current_subscription.canceled_at = timezone.now()
-                            self.customer.current_subscription.save()
+                        if Customer.allow_multiple_subscriptions:
+                            self.customer.sync_subscriptions()
                         else:
-                            self.customer.sync_current_subscription()
-                        """
-                        self.customer.sync_subscriptions()
+                            if self.kind == "customer.subscription.deleted":
+                                self.customer.current_subscription.status = Subscription.STATUS_CANCELLED
+                                self.customer.current_subscription.canceled_at = timezone.now()
+                                self.customer.current_subscription.save()
+                            else:
+                                self.customer.sync_current_subscription()
                 elif self.kind == "customer.deleted":
-                    if not self.customer:
-                        self.link_customer()
                     self.customer.purge()
                 self.send_signal()
                 self.processed = True
                 self.save()
-            except stripe.StripeError as e:
+            except stripe.StripeError as exc:
                 EventProcessingException.log(
-                    data=e.http_body,
-                    exception=e,
+                    data=exc.http_body,
+                    exception=exc,
                     event=self
                 )
                 webhook_processing_error.send(
                     sender=Event,
-                    data=e.http_body,
-                    exception=e
+                    data=exc.http_body,
+                    exception=exc
                 )
 
     def send_signal(self):
@@ -250,8 +229,10 @@ class Transfer(StripeObject):
 
     objects = TransferManager()
 
+    def __str__(self):
+        return "<amount={amount}, status={status}, stripe_id={stripe_id}>".format(amount=self.amount, status=self.status, stripe_id=self.stripe_id)
+
     def update_status(self):
-        # TODO - needs test
         self.status = stripe.Transfer.retrieve(self.stripe_id).status
         self.save()
 
@@ -280,6 +261,7 @@ class Transfer(StripeObject):
         for field in defaults:
             if field.endswith("fees") or field.endswith("gross"):
                 defaults[field] = defaults[field] / decimal.Decimal("100")
+
         if event.kind == "transfer.paid":
             defaults.update({"event": event})
             obj, created = Transfer.objects.get_or_create(
@@ -292,6 +274,7 @@ class Transfer(StripeObject):
                 event=event,
                 defaults=defaults
             )
+
         if created:
             for fee in transfer["summary"]["charge_fee_details"]:
                 obj.charge_fee_details.create(
@@ -303,8 +286,8 @@ class Transfer(StripeObject):
         else:
             obj.status = transfer["status"]
             obj.save()
+
         if event.kind == "transfer.updated":
-            # TODO - needs test
             obj.update_status()
 
 
@@ -318,12 +301,12 @@ class TransferChargeFee(TimeStampedModel):
 
 @python_2_unicode_compatible
 class Customer(StripeObject):
-    # TODO - needs tests
-
     subscriber = models.OneToOneField(getattr(settings, 'DJSTRIPE_SUBSCRIBER_MODEL', settings.AUTH_USER_MODEL), null=True)
     card_fingerprint = models.CharField(max_length=200, blank=True)
     card_last_4 = models.CharField(max_length=4, blank=True)
     card_kind = models.CharField(max_length=50, blank=True)
+    card_exp_month = models.PositiveIntegerField(blank=True, null=True)
+    card_exp_year = models.PositiveIntegerField(blank=True, null=True)
     date_purged = models.DateTimeField(null=True, editable=False)
 
     objects = CustomerManager()
@@ -332,7 +315,7 @@ class Customer(StripeObject):
     retain_canceled_subscriptions = djstripe_settings.RETAIN_CANCELED_SUBSCRIPTIONS
 
     def __str__(self):
-        return unicode(self.subscriber)
+        return "<{subscriber}, stripe_id={stripe_id}>".format(subscriber=smart_text(self.subscriber), stripe_id=self.stripe_id)
 
     @property
     def stripe_customer(self):
@@ -341,8 +324,8 @@ class Customer(StripeObject):
     def purge(self):
         try:
             self.stripe_customer.delete()
-        except stripe.InvalidRequestError as e:
-            if str(e).startswith("No such customer:"):
+        except stripe.InvalidRequestError as exc:
+            if str(exc).startswith("No such customer:"):
                 # The exception was thrown because the stripe customer was already
                 # deleted on the stripe side, ignore the exception
                 pass
@@ -353,6 +336,8 @@ class Customer(StripeObject):
         self.card_fingerprint = ""
         self.card_last_4 = ""
         self.card_kind = ""
+        self.card_exp_month = None
+        self.card_exp_year = None
         self.date_purged = timezone.now()
         self.save()
 
@@ -455,8 +440,7 @@ class Customer(StripeObject):
         return subscription
 
     def cancel(self, at_period_end=True):
-        # TODO - add deprecation warning and test
-        """ Adapter method to preserve usage of previous API """
+        warnings.warn("Deprecated - Use ``cancel_subscription`` instead. This method will be removed in dj-stripe 1.0.", DeprecationWarning)
         return self.cancel_subscription(at_period_end=at_period_end)
 
     @classmethod
@@ -487,6 +471,8 @@ class Customer(StripeObject):
         self.card_fingerprint = cu.active_card.fingerprint
         self.card_last_4 = cu.active_card.last4
         self.card_kind = cu.active_card.type
+        self.card_exp_month = cu.active_card.exp_month
+        self.card_exp_year = cu.active_card.exp_year
         self.save()
         card_changed.send(sender=self, stripe_response=cu)
 
@@ -495,9 +481,9 @@ class Customer(StripeObject):
         for inv in self.invoices.filter(paid=False, closed=False):
             try:
                 inv.retry()  # Always retry unpaid invoices
-            except stripe.error.InvalidRequestError as error:
-                if error.message != "Invoice is already paid":
-                    raise error
+            except stripe.InvalidRequestError as exc:
+                if str(exc) != "Invoice is already paid":
+                    raise exc
 
     def send_invoice(self):
         try:
@@ -513,6 +499,8 @@ class Customer(StripeObject):
             self.card_fingerprint = cu.active_card.fingerprint
             self.card_last_4 = cu.active_card.last4
             self.card_kind = cu.active_card.type
+            self.card_exp_month = cu.active_card.exp_month
+            self.card_exp_year = cu.active_card.exp_year
             self.save()
 
     def sync_invoices(self, cu=None, **kwargs):
@@ -766,8 +754,6 @@ class Customer(StripeObject):
 
 
 class Subscription(StripeObject):
-    # TODO - needs tests
-
     STATUS_TRIALING = "trialing"
     STATUS_ACTIVE = "active"
     STATUS_PAST_DUE = "past_due"
@@ -825,10 +811,29 @@ class Subscription(StripeObject):
 
         return True
 
+    def extend(self, delta):
+        if delta.total_seconds() < 0:
+            raise ValueError("delta should be a positive timedelta.")
+
+        period_end = None
+
+        if self.trial_end is not None and \
+           self.trial_end > timezone.now():
+            period_end = self.trial_end
+        else:
+            period_end = self.current_period_end
+
+        period_end += delta
+
+        self.customer.stripe_customer.update_subscription(
+            prorate=False,
+            trial_end=period_end,
+        )
+
+        self.customer.sync_current_subscription()
+
 
 class Invoice(StripeObject):
-    # TODO - needs tests
-
     customer = models.ForeignKey(Customer, related_name="invoices")
     attempted = models.NullBooleanField()
     attempts = models.PositiveIntegerField(null=True)
@@ -843,6 +848,9 @@ class Invoice(StripeObject):
 
     class Meta:
         ordering = ["-date"]
+
+    def __str__(self):
+        return "<total={total}, paid={paid}, stripe_id={stripe_id}>".format(total=self.total, paid=smart_text(self.paid), stripe_id=self.stripe_id)
 
     def retry(self):
         if not self.paid and not self.closed:
@@ -954,8 +962,9 @@ class Invoice(StripeObject):
 
 
 class InvoiceItem(TimeStampedModel):
-    """ Not inherited from StripeObject because there can be multiple invoice
-        items for a single stripe_id.
+    """
+    Not inherited from StripeObject because there can be multiple invoice
+    items for a single stripe_id.
     """
 
     stripe_id = models.CharField(max_length=50)
@@ -970,13 +979,14 @@ class InvoiceItem(TimeStampedModel):
     plan = models.CharField(max_length=100, null=True, blank=True)
     quantity = models.IntegerField(null=True)
 
+    def __str__(self):
+        return "<amount={amount}, plan={plan}, stripe_id={stripe_id}>".format(amount=self.amount, plan=smart_text(self.plan), stripe_id=self.stripe_id)
+
     def plan_display(self):
-        # TODO - needs test
         return djstripe_settings.PAYMENTS_PLANS[self.plan]["name"]
 
 
 class Charge(StripeObject):
-
     customer = models.ForeignKey(Customer, related_name="charges")
     invoice = models.ForeignKey(Invoice, null=True, related_name="charges")
     card_last_4 = models.CharField(max_length=4, blank=True)
@@ -993,6 +1003,9 @@ class Charge(StripeObject):
     charge_created = models.DateTimeField(null=True, blank=True)
 
     objects = ChargeManager()
+
+    def __str__(self):
+        return "<amount={amount}, paid={paid}, stripe_id={stripe_id}>".format(amount=self.amount, paid=smart_text(self.paid), stripe_id=self.stripe_id)
 
     def calculate_refund_amount(self, amount=None):
         eligible_to_refund = self.amount - (self.amount_refunded or 0)
@@ -1023,7 +1036,6 @@ class Charge(StripeObject):
         obj, _ = customer.charges.get_or_create(stripe_id=data["id"])
         invoice_id = data.get("invoice", None)
         if obj.customer.invoices.filter(stripe_id=invoice_id).exists():
-            # TODO - needs test
             obj.invoice = obj.customer.invoices.get(stripe_id=invoice_id)
         obj.card_last_4 = data["card"]["last4"]
         obj.card_kind = data["card"]["type"]
@@ -1035,7 +1047,6 @@ class Charge(StripeObject):
         obj.disputed = data["dispute"] is not None
         obj.charge_created = convert_tstamp(data, "created")
         if data.get("description"):
-            # TODO - needs test
             obj.description = data["description"]
         if data.get("amount_refunded"):
             obj.amount_refunded = (data["amount_refunded"] / decimal.Decimal("100"))
@@ -1096,7 +1107,7 @@ class Plan(StripeObject):
     trial_period_days = models.IntegerField(null=True)
 
     def __str__(self):
-        return self.name
+        return "<{name}, stripe_id={stripe_id}>".format(name=smart_text(self.name), stripe_id=self.stripe_id)
 
     @classmethod
     def create(cls, metadata={}, **kwargs):
@@ -1150,5 +1161,4 @@ class Plan(StripeObject):
     @property
     def stripe_plan(self):
         """Return the plan data from Stripe."""
-        # TODO - needs test
         return stripe.Plan.retrieve(self.stripe_id)